build(jenkins): remove image only docker push stage is success

+++ Jenkinsfile
@@ -113,13 +113,15 @@ pipeline {
     post {
         always {
             script {
-                sh """
-                    docker rmi ${IMAGE_NAME}:${IMAGE_TAG} 2>/dev/null || echo 'Image ${IMAGE_NAME}:${IMAGE_TAG} not found'
-                    docker rmi ${IMAGE_NAME}:latest 2>/dev/null || echo 'Image ${IMAGE_NAME}:latest not found'
+                if (currentBuild.rawBuild.getPreviousBuild()?.getAction(org.jenkinsci.plugins.workflow.actions.StageAction)?.getStageFlowNodes()?.find { it.displayName == 'Build & Push Docker Image' }?.isSuccess()) {
+                    sh """
+                        docker rmi ${IMAGE_NAME}:${IMAGE_TAG} 2>/dev/null || echo 'Image ${IMAGE_NAME}:${IMAGE_TAG} not found'
+                        docker rmi ${IMAGE_NAME}:latest 2>/dev/null || echo 'Image ${IMAGE_NAME}:latest not found'
 
-                    # Clean up unused Docker resources
-                    docker system prune -f 2>/dev/null || true
-                """
+                        # Clean up unused Docker resources
+                        docker system prune -f 2>/dev/null || true
+                    """
+                }
             }
         }
